// src/common/enums/notification-type.enum.ts
export enum NotificationType {
  RegistrationConfirmation = 'registration_confirmation',
  PaymentConfirmation = 'payment_confirmation',
  EventReminder = 'event_reminder',
  WaitlistConfirmation = 'waitlist_confirmation',
  SpotAvailable = 'spot_available',
  CancellationConfirmation = 'cancellation_confirmation',
  EventChanged = 'event_changed',
}